% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6-package.R
\docType{package}
\name{cmip6-package}
\alias{cmip6}
\alias{cmip6-package}
\title{cmip6: Functions for Downloading the NASA NEX-GDDP-CMIP6 Downscaled Climate Projections}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Kyle Bocinsky \email{bocinsky@gmail.com} (\href{https://orcid.org/0000-0003-1862-3428}{ORCID})

}
\keyword{internal}
