% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_dl_aws.R
\name{cmip6_dl_aws}
\alias{cmip6_dl_aws}
\title{Download NASA NEX-GDDP-CMIP6 data from Amazon Web Services}
\usage{
cmip6_dl_aws(plan, outdir, workers, force)
}
\arguments{
\item{force}{}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble::tbl_df}} listing the downloaded data.
}
\description{
Download NASA NEX-GDDP-CMIP6 data from Amazon Web Services
}
