% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_dl_tds.R
\name{cmip6_dl_tds}
\alias{cmip6_dl_tds}
\title{Download NASA NEX-GDDP-CMIP6 data from the NCCS THREDDS Data Catalog}
\usage{
cmip6_dl_tds(aoi, plan, outdir, workers, force)
}
\arguments{
\item{aoi}{Optional. An \code{\link{sf}} object specifying an area of interest.
If omitted, data for the entire globe are downloaded.}

\item{plan}{A \code{\link[tibble:tbl_df-class]{tibble::tbl_df}} filtering data from \code{\link{cmip6_ls}} that
contains the data to be downloaded. Generated by \code{\link{cmip6_dl}}.}

\item{outdir}{The directory to which to write the downloaded data. Defaults
to the current working directory.}

\item{workers}{Parallelization is supported by default with 10 workers.
Specify an integer value of more or fewer workers if desired. Note that
specifying too many workers may cause download requests to be rejected
by the NASA NEX-GDDP-CMIP6 host servers.}

\item{force}{Whether to re-download data that already exists on disk.
Defaults to \code{FALSE}.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble::tbl_df}} listing the downloaded data.
}
\description{
Use this function if you want to download NASA NEX-GDDP-CMIP6 data for a
specific area of interest (\code{aoi}). Called via \code{\link{cmip6_dl}} when an
\code{aoi} is specified.
}
