% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_ls.R
\name{cmip6_ls}
\alias{cmip6_ls}
\title{List the available NASA NEX-GDDP-CMIP6 data}
\usage{
cmip6_ls(latest = TRUE)
}
\arguments{
\item{latest}{Whether to remove legacy versions of the data where
new ones exist. Defaults to TRUE.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble::tbl_df}} listing the available data.
}
\description{
List the available NASA NEX-GDDP-CMIP6 data
}
\examples{
cmip6_ls()
}
