% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download a file using httr2}
\usage{
download_file(url, out.path, force)
}
\arguments{
\item{url}{A URL to download.}

\item{out.path}{A path to which to write the downloaded file.}

\item{force}{Whether to overwrite the file if it exists.}
}
\value{
The path of the downloaded file.
}
\description{
Download a file using httr2
}
