% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cmip6.R
\name{install_cmip6}
\alias{install_cmip6}
\title{Install chelsa cmip6 package from conda}
\usage{
install_cmip6(
  ...,
  envname = "r-cmip6",
  new_env = identical(envname, "r-cmip6")
)
}
\arguments{
\item{...}{Passing additional parameters}

\item{envname}{The name of a conda environment attached to this package}

\item{new_env}{Set a new environment where to install this package}
}
\description{
Install chelsa cmip6 package from conda
}
