% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6_dl.R
\name{cmip6_dl}
\alias{cmip6_dl}
\title{Title}
\usage{
cmip6_dl(
  outdir,
  aoi = NULL,
  models = NULL,
  scenarios = NULL,
  elements = NULL,
  years = NULL,
  latest = TRUE,
  workers = 10,
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\description{
Title
}
