% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_rotate}
\alias{st_rotate}
\title{Change a simple feature from [-180,180) longitude to [0,360) longitude}
\usage{
st_rotate(x)
}
\arguments{
\item{x}{An \code{\link{sf}} object.}
}
\value{
A rotate \code{\link{sf}} object.
}
\description{
Change a simple feature from [-180,180) longitude to [0,360) longitude
}
